move exception class to the exception package
