+++ raffaelliscandiffio.shop-totem/src/it/java/com/github/raffaelliscandiffio/repository/service/ServiceMongoIT.java
@@ -14,10 +14,10 @@ import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 
+import com.github.raffaelliscandiffio.exception.TransactionException;
 import com.github.raffaelliscandiffio.model.Product;
 import com.github.raffaelliscandiffio.model.Stock;
 import com.github.raffaelliscandiffio.service.ShoppingService;
-import com.github.raffaelliscandiffio.transaction.TransactionException;
 import com.github.raffaelliscandiffio.transaction.mongo.TransactionManagerMongo;
 import com.mongodb.client.ClientSession;
 import com.mongodb.client.MongoClient;
