+++ raffaelliscandiffio.shop-totem/src/main/java/com/github/raffaelliscandiffio/exception/TransactionException.java
@@ -1,4 +1,4 @@
-package com.github.raffaelliscandiffio.transaction;
+package com.github.raffaelliscandiffio.exception;
 
 public class TransactionException extends RuntimeException {
