+++ raffaelliscandiffio.shop-totem/src/main/java/com/github/raffaelliscandiffio/transaction/mongo/TransactionManagerMongo.java
@@ -1,11 +1,11 @@
 package com.github.raffaelliscandiffio.transaction.mongo;
 
+import com.github.raffaelliscandiffio.exception.TransactionException;
 import com.github.raffaelliscandiffio.repository.mongo.OrderItemMongoRepository;
 import com.github.raffaelliscandiffio.repository.mongo.OrderMongoRepository;
 import com.github.raffaelliscandiffio.repository.mongo.ProductMongoRepository;
 import com.github.raffaelliscandiffio.repository.mongo.StockMongoRepository;
 import com.github.raffaelliscandiffio.transaction.TransactionCode;
-import com.github.raffaelliscandiffio.transaction.TransactionException;
 import com.github.raffaelliscandiffio.transaction.TransactionManager;
 import com.mongodb.client.MongoClient;
 import com.mongodb.client.ClientSession;
