+++ raffaelliscandiffio.shop-totem/src/main/java/com/github/raffaelliscandiffio/transaction/mysql/TransactionManagerMySql.java
@@ -2,12 +2,12 @@ package com.github.raffaelliscandiffio.transaction.mysql;
 
 import javax.persistence.EntityManager;
 
+import com.github.raffaelliscandiffio.exception.TransactionException;
 import com.github.raffaelliscandiffio.repository.mysql.OrderItemMySqlRepository;
 import com.github.raffaelliscandiffio.repository.mysql.OrderMySqlRepository;
 import com.github.raffaelliscandiffio.repository.mysql.ProductMySqlRepository;
 import com.github.raffaelliscandiffio.repository.mysql.StockMySqlRepository;
 import com.github.raffaelliscandiffio.transaction.TransactionCode;
-import com.github.raffaelliscandiffio.transaction.TransactionException;
 import com.github.raffaelliscandiffio.transaction.TransactionManager;
 
 public class TransactionManagerMySql implements TransactionManager {
